+++ include/RMRT/Objects/Sphere.h
@@ -9,18 +9,18 @@ namespace rmrt {
 	{
 	public:
 		Sphere();
-		Sphere(Point3 center, double radius, std::shared_ptr<Material> material);
+		Sphere(Point3 center, float radius, std::shared_ptr<Material> material);
 		~Sphere();
 
 		void SetCenter(const Point3& center);
 		const Point3& GetCenter() const;
-		void SetRadius(const double& radius);
-		const double& GetRadius() const;
-		virtual bool Hit(const Ray& ray, double tMin, double tMax, HitRecord& record) const override;
+		void SetRadius(const float& radius);
+		const float& GetRadius() const;
+		virtual bool Hit(const Ray& ray, float tMin, float tMax, HitRecord& record) const override;
 
 	private:
 		Point3 m_center;
-		double m_radius;
+		float m_radius;
 		std::shared_ptr<Material> m_materialPtr;
 	};
 }
